new integration tests structure
