+++ integration/openid/src/test/resources/simplelogger.properties
@@ -1,0 +1,1 @@
+org.slf4j.simpleLogger.log.org.eclipse.jetty=warn
