+++ integration/parent/src/main/resources/META-INF/microprofile-config.properties
