+++ integration/parent/src/main/resources/metadata.json
