+++ integration/parent/src/main/resources/scenarios.ttl
