+++ integration/parent/src/main/resources/simplelogger.properties
