+++ integration/parent/src/main/resources/uma2-configuration.json
