+++ integration/src/site/site.xml
@@ -7,7 +7,7 @@
   </bannerRight>
 
   <body>
-    <menu name="Module: Examples" inherit="top">
+    <menu name="Module: Integration Tests" inherit="top">
         <item name="Summary" href="summary.html"/>
         <item name="Dependencies" href="dependencies.html"/>
         <item name="Plugins" href="plugins.html"/>
